+++ tests/run/tryfinally.pyx
@@ -3,6 +3,12 @@
 
 cimport cython
 
+try:
+    next
+except NameError:
+    def next(it): return it.next()
+
+
 def finally_except():
     """
     >>> try:
@@ -161,3 +167,85 @@ def empty_try_in_except_raise(raise_in_finally):
             if raise_in_finally:
                 raise TypeError('OLA')
         raise
+
+
+def try_all_cases(x):
+    """
+    >>> try_all_cases(None)
+    2
+    >>> try_all_cases('break')
+    4
+    >>> try_all_cases('raise')
+    Traceback (most recent call last):
+    ValueError
+    >>> try_all_cases('return')
+    3
+    >>> try_all_cases('tryraise')
+    Traceback (most recent call last):
+    TypeError
+    >>> try_all_cases('trybreak')
+    4
+    """
+    for i in range(3):
+        try:
+            if i == 0:
+                pass
+            elif i == 1:
+                continue
+            elif x == 'trybreak':
+                break
+            elif x == 'tryraise':
+                raise TypeError()
+            else:
+                return 2
+        finally:
+            if x == 'raise':
+                raise ValueError()
+            elif x == 'break':
+                break
+            elif x == 'return':
+                return 3
+    return 4
+
+
+def finally_yield(x):
+    """
+    >>> g = finally_yield(None)
+    >>> next(g)  # 1
+    1
+    >>> next(g)  # 2
+    1
+    >>> next(g)  # 3
+    Traceback (most recent call last):
+    StopIteration
+
+    >>> g = finally_yield('raise')
+    >>> next(g)  # raise 1
+    1
+    >>> next(g)  # raise 2
+    1
+    >>> next(g)  # raise 3
+    Traceback (most recent call last):
+    TypeError
+
+    >>> g = finally_yield('break')
+    >>> next(g)   # break 1
+    1
+    >>> next(g)   # break 2
+    1
+    >>> next(g)   # break 3
+    Traceback (most recent call last):
+    StopIteration
+    """
+    for i in range(3):
+        try:
+            if i == 0:
+                continue
+            elif x == 'raise':
+                raise TypeError()
+            elif x == 'break':
+                break
+            else:
+                return
+        finally:
+            yield 1